requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.51
langchain-google-genai==2.1.2
langchain-text-splitters==0.3.8
langsmith==0.3.24
orjson==3.10.16
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic_core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
